third_party/blink/web_tests/platform/generic/fast/events/onbeforeunload-focused-iframe-expected.txt

CONSOLE ERROR: Blocked alert('beforeUnload') during beforeunload.
New window with beforeUnload handler.
You should see an alert with "beforeUnload" if you try to close the window.


--------
Frame: 'focusedFrame'
--------
This frame should get the focus. 
